controller/app/requirements.txt
uvicorn==0.27.0
pydantic==2.5.3
fastapi==0.109.0
aiomqtt==2.0.0
roboticstoolbox-python==1.1.0
scipy<1.12.0
matplotlib==3.8.2
redis==5.0.1
